backend/pyproject.toml
[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.mypy]
files = ["app", "data", "dependencies" , "tests"]
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true

[tool.flake8]
max-line-length = 120
ignore = ["E203","W501","W503"]

[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A prototype asset manamgement"
authors = ["kcw2297 <kcw2371@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.6.0"
anyio = "4.2.0"
appnope = "0.1.4"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asttokens = "2.4.1"
async-lru = "2.0.4"
async-timeout = "4.0.3"
attrs = "23.2.0"
authlib = "1.3.0"
babel = "2.14.0"
bcrypt = "4.1.2"
beanie = "1.25.0"
beautifulsoup4 = "4.12.3"
black = "24.3.0"
bleach = "6.1.0"
bs4 = "0.0.2"
build = "1.1.1"
cachecontrol = "0.14.0"
cachetools = "5.3.3"
certifi = "2024.2.2"
cffi = "1.16.0"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
comm = "0.2.1"
crashtest = "0.4.1"
cryptography = "42.0.3"
debugpy = "1.8.1"
decorator = "5.1.1"
defusedxml = "0.7.1"
distlib = "0.3.8"
dnspython = "2.6.0"
dulwich = "0.21.7"
email-validator = "2.1.0.post1"
executing = "2.0.1"
fastapi = "0.109.2"
fastapi-users = "12.1.3"
fastapi-users-db-beanie = "3.0.0"
fastapi-users-db-sqlalchemy = "6.0.1"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
finance-datareader = "0.9.70"
fqdn = "1.5.1"
google-auth = "2.28.2"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.3"
httpx = "0.26.0"
httpx-oauth = "0.13.2"
identify = "2.5.35"
idna = "3.6"
importlib-metadata = "7.0.2"
iniconfig = "2.0.0"
installer = "0.7.0"
ipykernel = "6.29.2"
ipython = "8.22.1"
isoduration = "20.11.0"
isort = "5.13.2"
itsdangerous = "2.1.2"
jaraco-classes = "3.3.1"
jedi = "0.19.1"
jinja2 = "3.1.3"
json5 = "0.9.17"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
jupyter-events = "0.9.0"
jupyter-lsp = "2.2.2"
jupyter-client = "8.6.0"
jupyter-core = "5.7.1"
jupyter-server = "2.12.5"
jupyter-server-terminals = "0.5.2"
jupyterlab = "4.1.2"
jupyterlab-pygments = "0.3.0"
jupyterlab-server = "2.25.3"
keyring = "24.3.1"
lazy-model = "0.2.0"
lxml = "5.1.0"
makefun = "1.15.2"
markupsafe = "2.1.5"
matplotlib-inline = "0.1.6"
mistune = "3.0.2"
more-itertools = "10.2.0"
motor = "3.3.2"
msgpack = "1.0.8"
mypy = "1.9.0"
mypy-extensions = "1.0.0"
nbclient = "0.9.0"
nbconvert = "7.16.1"
nbformat = "5.9.2"
nest-asyncio = "1.6.0"
nodeenv = "1.8.0"
notebook = "7.1.0"
notebook-shim = "0.2.4"
numpy = "1.26.4"
outcome = "1.3.0.post0"
overrides = "7.7.0"
packaging = "23.2"
pandas = "2.2.1"
pandocfilters = "1.5.1"
parso = "0.8.3"
passlib = "1.7.4"
pathspec = "0.12.1"
pexpect = "4.9.0"
pkginfo = "1.10.0"
platformdirs = "4.2.0"
plotly = "5.19.0"
pluggy = "1.4.0"
poetry = "1.8.2"
poetry-core = "1.9.0"
poetry-plugin-export = "1.7.0"
pre-commit = "3.6.2"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.43"
psutil = "5.9.8"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
pygments = "2.17.2"
pyjwt = "2.8.0"
pymongo = "4.6.1"
pyproject-hooks = "1.0.0"
pysocks = "1.7.1"
pytest = "7.4.4"
pytest-asyncio = "0.23.5"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
python-json-logger = "2.0.7"
python-multipart = "0.0.7"
pytz = "2024.1"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
rapidfuzz = "3.6.2"
redis = "4.6.0"
referencing = "0.33.0"
requests = "2.31.0"
requests-file = "2.0.0"
requests-toolbelt = "1.0.0"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rpds-py = "0.18.0"
rsa = "4.9"
selenium = "4.18.1"
send2trash = "1.8.2"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
soupsieve = "2.5"
sqlalchemy = "2.0.24"
stack-data = "0.6.3"
starlette = "0.36.3"
tenacity = "8.2.3"
terminado = "0.18.0"
tinycss2 = "1.2.1"
toml = "0.10.2"
tomlkit = "0.12.4"
tornado = "6.4"
tqdm = "4.66.2"
traitlets = "5.14.1"
trio = "0.24.0"
trio-websocket = "0.11.1"
trove-classifiers = "2024.3.3"
types-python-dateutil = "2.8.19.20240106"
typing-extensions = "4.9.0"
tzdata = "2024.1"
uri-template = "1.3.0"
urllib3 = "2.2.1"
uvicorn = "0.27.0"
virtualenv = "20.25.1"
wcwidth = "0.2.13"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.7.0"
wsproto = "1.2.0"
xattr = "1.1.0"
zipp = "3.18.1"
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"
types-redis = "4.6.0.20240311"
pytest-mock = "^3.14.0"
openpyxl = "^3.1.2"
websockets = "^12.0"
crypto = "^1.4.1"
pypdf2 = "^3.0.1"
pandas-datareader = "^0.10.0"
aiomysql = "^0.2.0"
alembic = "^1.13.1"
aioredis = "^2.0.1"
aiohttp = "^3.9.5"
boto3 = "^1.34.113"
mysql-connector-python = "^8.4.0"
yfinance = "^0.2.40"
nasdaq-data-link = "^1.0.4"
aiosqlite = "^0.20.0"
polygon = "^1.2.0"
polygon-api-client = "^1.14.1"
fakeredis = "^2.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
